MAINT: Add basic checks for minimum terminal size
These are currently only applied when the window is first created,
though. It would be better to have this check always occurring during
gameplay and allow the user to increase the size again instead of just
crashing if they inadvertently made the terminal too small in the middle
of a game.