BUG: Add workaround to ensure main help menu is displayed
It seems, at least on mac, that if you place_forget a DismissiblePopup
and then replace it, the contents will be empty until the mouse passes
in or out of the widget, which must force tkinter to think it needs to
refresh the widget. Here, we use update_idletasks to force tkinter to
flush its event queue.