MAINT: Create magic key command context
This context will eventually be switched to by the primary interface
context when the user attempts to walk through a permanently locked door
while having 1 or more magic keys in their inventory. The model's
"move_adventurer" method should return some value that indicates that
this should be done...yet to be finalized.